+++ src/storage/indexmapped.rs
@@ -31,8 +31,9 @@ where
     }
 
     fn insert_new_cell(&self, cell: Cell, key: K) {
-        self.key_to_cell.insert(key.clone(), cell).ok();
-        self.cell_to_key.insert(cell, (key, None)).ok();
+        // key_to_cell must be written last to avoid data races
+        self.cell_to_key.insert(cell, (key.clone(), None)).ok();
+        self.key_to_cell.insert(key, cell).ok();
     }
 
     fn try_get_input(&self, cell: Cell) -> Option<K> {